% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installer.R
\name{install_app}
\alias{install_app}
\title{Download and install Wilson Apps}
\usage{
install_app(
  location,
  remove_data = FALSE,
  start_after_install = FALSE,
  app_name = "wilson-basic",
  repository = "https://github.molgen.mpg.de/loosolab/wilson-apps"
)
}
\arguments{
\item{location}{Where the app should be installed.}

\item{remove_data}{If TRUE demo data will be deleted.}

\item{start_after_install}{Start the app when done installing.}

\item{app_name}{Select app to install.}

\item{repository}{Link to the repository that holds the apps.}
}
\description{
Download and install Wilson Apps
}
\details{
Will create a folder named after parameter app_name.
}
