% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{log_message}
\alias{log_message}
\title{logger message convenience function}
\usage{
log_message(
  message,
  level = c("DEBUG", "INFO", "WARN", "ERROR", "FATAL"),
  token = NULL
)
}
\arguments{
\item{message}{String of message to be written in log. See \code{\link[lgr]{Logger}}.}

\item{level}{Set priority level of the message (number or character). See \code{\link[lgr]{Logger}}.}

\item{token}{Use token bound to this identifier.}
}
\description{
logger message convenience function
}
\details{
Does nothing if logger doesn't exist.
}
