% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_logger}
\alias{set_logger}
\title{set a logger used within the package}
\usage{
set_logger(logger, token = NULL)
}
\arguments{
\item{logger}{A logger object, see \code{\link{create_logger}} or \code{\link[lgr]{get_logger}}. NULL to disable logging.}

\item{token}{Set a unique identifier for this logger.}
}
\description{
set a logger used within the package
}
\details{
This function will save each logger in the wilson.globals environment. Each logger is stored by the name 'logger'[token] (e.g. 'logger6b821824b0b53b1a3e8f531a34d0d6e6').

Use onSessionEnded to clean up after logging. See \code{\link[shiny]{onFlush}}.
}
